chore: fix circular dependency - move ServiceType from ScanningStrategyD
