+++ src/detectors/IScanningStrategy.ts
@@ -1,0 +1,13 @@
+export enum ServiceType {
+  github = 'github',
+  bitbucket = 'bitbucket',
+  gitlab = 'gitlab',
+  git = 'git',
+  local = 'local',
+}
+
+export enum AccessType {
+  private = 'private',
+  public = 'public',
+  unknown = 'unknown',
+}
